+++ nms-shared/adapter/legacy/AbstractNmsAdapter.java
@@ -9,6 +9,7 @@ import net.opmasterleo.packetuxui.nms.NmsUnsupportedException;
 import net.opmasterleo.packetuxui.nms.PacketClassifier;
 import net.opmasterleo.packetuxui.nms.PipelineBridge;
 import net.opmasterleo.packetuxui.nms.ProtocolVersions;
+import net.opmasterleo.packetuxui.nms.SignPacketBridge;
 
 public abstract class AbstractNmsAdapter implements NmsAdapter {
 
@@ -19,6 +20,7 @@ public abstract class AbstractNmsAdapter implements NmsAdapter {
     private final SharedMenuPacketBridge packets;
     private final SharedPipelineBridge pipeline;
     private final SharedPacketClassifier classifier;
+    private final SharedSignPackets signs;
 
     protected AbstractNmsAdapter(String bucketId, String nmsPackage, int minProtocol, int maxProtocol) {
         String serverPackage = Bukkit.getServer().getClass().getPackage().getName();
@@ -34,6 +36,7 @@ public abstract class AbstractNmsAdapter implements NmsAdapter {
         this.packets = new SharedMenuPacketBridge(items);
         this.pipeline = new SharedPipelineBridge();
         this.classifier = new SharedPacketClassifier(items);
+        this.signs = new SharedSignPackets();
     }
 
     @Override
@@ -69,5 +72,10 @@ public abstract class AbstractNmsAdapter implements NmsAdapter {
     @Override
     public ItemBridge items() {
         return items;
+    }
+
+    @Override
+    public SignPacketBridge signs() {
+        return signs;
     }
 }